requirements.txt
#pip install -r requirements.txt
# core dependencies
Flask>=3.0.0
gunicorn>=22.0.0
requests>=2.31.0
python-dotenv>=1.0.0
xlrd==1.2.0
authlib>=1.3.0
onnxruntime>=1.18.0
Pillow>=10.0.0
XlsxWriter>=3.2.0
minio>=7.1.0
boto3>=1.28.0
#outlier detection model compatibility
joblib==1.4.2
matplotlib==3.8.0
numpy==1.24.3     
pandas==2.3.3
pyod==2.0.3
scikit-learn==1.2.2  
xgboost==2.0.3
openpyxl== 3.1.5



